feat: refactor endpoint retrieval and enhance daily challenge query handling
